# openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Generated by InducOapi
  version: v1
servers:
- url: /
paths:
  /auth:
    post:
      operationId: ldapauth_post
      requestBody:
        $ref: '#/components/requestBodies/LDAPAuth'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ldapAuth_response'
          description: this a default description
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /verify:
    post:
      operationId: verifytoken_post
      requestBody:
        $ref: '#/components/requestBodies/VerifyToken'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/verifyToken_response'
          description: this a default description
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /isalive:
    get:
      operationId: get_isalive
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetIsAlive_response'
          description: this a default description
      x-openapi-router-controller: openapi_server.controllers.default_controller
components:
  requestBodies:
    LDAPAuth:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LDAPAuth'
    VerifyToken:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/VerifyToken'
  schemas:
    LDAPAuth:
      example:
        password: Pa$$w0rd
        username: jdupont
      properties:
        password:
          example: Pa$$w0rd
          type: string
        username:
          example: jdupont
          type: string
      type: object
    GetIsAlive_response_isalive:
      properties:
        uri:
          example: /isalive
          type: string
        title:
          example: IsAlive
          type: string
        description:
          example: The REST API is on air for GitInterface API
          type: string
      type: object
    GetIsAlive_response:
      example:
        isalive:
        - uri: /isalive
          title: IsAlive
          description: The REST API is on air for GitInterface API
        - uri: /isalive
          title: Version
          description: AI4IA REST API v1.0 (c) AI for Infrastructure Automation - Toulouse
      properties:
        isalive:
          items:
            $ref: '#/components/schemas/GetIsAlive_response_isalive'
          type: array
      type: object
    VerifyToken:
      example:
        auth_token: qijdoijoijqsoidjauzhiauzyeiuvcjw;vbnbkjj
      properties:
        auth_token:
          example: qijdoijoijqsoidjauzhiauzyeiuvcjw;vbnbkjj
          type: string
      type: object
    ldapAuth_response:
      example:
        cr: "001"
        result: ijoaijdoijaodijosaijduhoiuxhicwuyflkandoknon
      properties:
        cr:
          example: "001"
          type: string
        result:
          example: ijoaijdoijaodijosaijduhoiuxhicwuyflkandoknon
          type: string
      type: object
    verifyToken_response:
      example:
        cr: "001"
        result: Token OK
        username: jdupont
      properties:
        cr:
          example: "001"
          type: string
        result:
          example: Token OK
          type: string
        username:
          example: jdupont
          type: string
      type: object